requirements.txt
# requirements.txt

# Telegram Bot Framework
python-telegram-bot[job-queue]

# Web Scraping & HTTP Requests
requests
beautifulsoup4

# Environment Variable Management
python-dotenv

# Caching Utility
cachetools